# perf: Optimize imports to make importing copick quicker and keep CLI snappy