--- a/src/copick/cli/add.py
+++ b/src/copick/cli/add.py
@@ -5,10 +5,7 @@ from typing import Tuple
 import click
 import tqdm
 
-import copick
 from copick.cli.util import add_config_option, add_create_overwrite_options, add_debug_option
-from copick.ops.add import _add_tomogram_mrc, _add_tomogram_zarr, add_object, add_object_volume, add_segmentation
-from copick.util.formats import get_data_from_file, get_format_from_extension
 from copick.util.log import get_logger
 
 
@@ -108,6 +105,10 @@ def tomogram(
 
     PATH: Path to the tomogram file (MRC or Zarr format) or glob pattern.
     """
+    # Deferred imports for performance
+    import copick
+    from copick.ops.add import _add_tomogram_mrc, _add_tomogram_zarr
+
     logger = get_logger(__name__, debug=debug)
 
     # Get root
@@ -256,6 +257,10 @@ def segmentation(
 
     PATH: Path to the segmentation file (MRC or Zarr format) or glob pattern.
     """
+    # Deferred import for performance
+    import copick
+    from copick.ops.add import add_segmentation
+
     logger = get_logger(__name__, debug=debug)
 
     # Get root
@@ -419,6 +424,11 @@ def object(
     """
     Add a pickable object to the project configuration.
     """
+    # Deferred import for performance
+    import copick
+    from copick.ops.add import add_object
+    from copick.util.formats import get_data_from_file, get_format_from_extension
+
     logger = get_logger(__name__, debug=debug)
 
     # Get root
@@ -528,6 +538,11 @@ def object_volume(
     """
     Add volume data to an existing pickable object.
     """
+    # Deferred import for performance
+    import copick
+    from copick.ops.add import add_object_volume
+    from copick.util.formats import get_data_from_file, get_format_from_extension
+
     logger = get_logger(__name__, debug=debug)
 
     # Get root