--- a/src/copick/impl/filesystem.py
+++ b/src/copick/impl/filesystem.py
@@ -31,7 +31,7 @@ from copick.models import (
 )
 from copick.util.log import get_logger
 
-# Dont import Geometry at runtime to keep CLI snappy
+# Don't import Geometry at runtime to keep CLI snappy
 if TYPE_CHECKING:
     from trimesh.parent import Geometry