--- a/src/copick/impl/overlay.py
+++ b/src/copick/impl/overlay.py
@@ -19,7 +19,7 @@ from copick.models import (
 )
 from copick.util.log import get_logger
 
-# Dont import Geometry at runtime to keep CLI snappy
+# Don't import Geometry at runtime to keep CLI snappy
 if TYPE_CHECKING:
     from trimesh.parent import Geometry