Several import optimizations to make importing copick and using the CLI snappier: 

#### 1. Defer imports for CLI tools until CLI calls. 
Moved all imports from `ops` and other places into the actual CLI calls.

#### 2. Defer import of trimesh
Trimesh imports from scipy, so all trimesh imports are deferred until `load()` methods are called. Trimesh imports for type checking updated to use `TYPE_CHECKING`.

#### 3. Defer import of OME-Zarr and scikit-image
All imports of ome_zarr.writer and scikit-image functions in `util/ome.py` are deferred to actual scaling or write calls. 

Timing: 

**Import**
copick-1.6.1: `import copick` -> 1.881 s
This PR: `import copick` -> 0.301 s

**CLI**
copick-1.6.1
```
time copick --version
copick --version  1.01s user 0.34s system 85% cpu 1.582 total
```

This PR: 
```
time copick --version
copick --version  0.20s user 0.10s system 65% cpu 0.470 total
```